pyproject.toml
[project]
name = "uulm-utils"
version = "0.1.0"
description = "Collection of helpers for Ulm University"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncclick>=8.1.8",
    "click>=8.1.8",
    "pandas>=2.2.3",
    "playwright>=1.51.0",
    "python-dotenv>=1.1.0",
    "questionary>=2.1.0",
]
[project.scripts]
uulm = "uulm_utils.main:cli"
